requirements.txt
mcp-python>=0.1.0
requests>=2.31.0
pytest>=7.4.0
python-dotenv>=1.0.0
pytest-asyncio>=0.23.0
numpy>=1.24.0 